--- .eslintrc
@@ -1,26 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "extends": [
-    "airbnb-base",
-    "prettier"
-  ],
-  "plugins": [
-    "prettier"
-  ],
-  "env": {
-    "jest": true
-  },
-  "rules": {
-    "no-underscore-dangle": 0,
-    "no-unused-vars": ["error", { "ignoreRestSiblings": true }],
-    "curly": "error",
-    "prettier/prettier": [
-      "error",
-      {
-        "singleQuote": true,
-        "semi": false,
-        "trailingComma": "all"
-      }
-    ]
-  }
-}
