+++ .eslintrc.json
@@ -11,6 +11,7 @@
     "@typescript-eslint/no-explicit-any": 0,
     "@typescript-eslint/explicit-module-boundary-types": 0,
     "@typescript-eslint/ban-ts-comment": 0,
+    "@typescript-eslint/no-empty-function": 0,
     "semi": ["error", "never"]
   }
 }
