+++ src/index.ts
@@ -49,7 +49,9 @@ class FetchableSmallFasta {
     const data = await this.data
     const entry = data.find(iter => iter.id === id)
     const length = end - start
-    if (!entry) throw new Error(`no sequence with id ${id} exists`)
+    if (!entry) {
+      throw new Error(`no sequence with id ${id} exists`)
+    }
     return entry.sequence.substr(start, length)
   }
